[DXP-1322] Rename `streamx manage` to `streamx manager`
