+++ core/src/main/java/dev/streamx/cli/BannerPrinter.java
@@ -2,7 +2,7 @@ package dev.streamx.cli;
 
 import static dev.streamx.cli.util.Output.print;
 
-import dev.streamx.cli.command.manage.ManageCommand;
+import dev.streamx.cli.command.manager.ManagerCommand;
 import dev.streamx.cli.command.run.RunCommand;
 import jakarta.enterprise.context.ApplicationScoped;
 import java.util.Set;
@@ -13,7 +13,7 @@ import picocli.CommandLine.ParseResult;
 public class BannerPrinter {
 
   private static final Set<String> COMMANDS_REQUIRING_PRINTING_BANNER =
-      Set.of(RunCommand.COMMAND_NAME, ManageCommand.COMMAND_NAME);
+      Set.of(RunCommand.COMMAND_NAME, ManagerCommand.COMMAND_NAME);
 
   private static final String BANNER = """
        ____  _                           __  __
