+++ core/src/main/java/dev/streamx/cli/StreamxCommand.java
@@ -2,7 +2,7 @@ package dev.streamx.cli;
 
 import dev.streamx.cli.command.ingestion.publish.PublishCommand;
 import dev.streamx.cli.command.ingestion.unpublish.UnpublishCommand;
-import dev.streamx.cli.command.manage.ManageCommand;
+import dev.streamx.cli.command.manager.ManagerCommand;
 import dev.streamx.cli.command.run.RunCommand;
 import dev.streamx.cli.config.ArgumentConfigSource;
 import dev.streamx.cli.config.validation.ConfigSourcesValidator;
@@ -24,7 +24,7 @@ import picocli.CommandLine.ParseResult;
 @TopCommand
 @Command(mixinStandardHelpOptions = true,
     name = "streamx",
-    subcommands = {RunCommand.class, ManageCommand.class,
+    subcommands = {RunCommand.class, ManagerCommand.class,
         PublishCommand.class, UnpublishCommand.class,
         HelpCommand.class},
     versionProvider = VersionProvider.class)
