--- core/src/main/java/dev/streamx/cli/command/manage/event/MeshManagerStarted.java
@@ -1,5 +1,0 @@
-package dev.streamx.cli.command.manage.event;
-
-public class MeshManagerStarted {
-
-}
