+++ core/src/main/java/dev/streamx/cli/command/manager/event/MeshManagerStarted.java
@@ -1,0 +1,5 @@
+package dev.streamx.cli.command.manager.event;
+
+public class MeshManagerStarted {
+
+}
