+++ core/src/main/resources/application.properties
@@ -8,7 +8,7 @@ smallrye.config.mapping.validate-unknown=false
 streamx.cli.settings.root-dir=${user.home}/.streamx
 streamx.runner.generated-keys.location=${user.home}/.streamx/config/generated
 
-streamx.manage.mesh-manager.image=ghcr.io/streamx-dev/streamx-mesh-manager/streamx-mesh-manager:0.0.1-jvm
+streamx.manager.mesh-manager.image=ghcr.io/streamx-dev/streamx-mesh-manager/streamx-mesh-manager:0.0.1-jvm
 
 %dev.streamx.cli.license.proceeding.enabled=false
 %test.streamx.cli.license.proceeding.enabled=false
